requirements.txt
torch>=2.0.0
python-chess
datasets
numpy
flask
